refactor: 행성 내 주민증 조회 시에도 대댓글 카운트 추가

+++ Api/src/main/java/com/dingdong/api/community/dto/CommunityIdCardsDto.java
@@ -32,9 +32,9 @@ public class CommunityIdCardsDto {
     private final List<String> keywordTitles;
 
     @Schema(description = "댓글 개수")
-    private final Long commentCount;
+    private final int commentCount;
 
-    public static CommunityIdCardsDto of(IdCard idCard, List<Keyword> keywords, Long commentCount) {
+    public static CommunityIdCardsDto of(IdCard idCard, List<Keyword> keywords, int commentCount) {
         return CommunityIdCardsDto.builder()
                 .idCardId(idCard.getId())
                 .nickname(idCard.getNickname())
